+++ mullvad-types/src/settings/migrations/v1.rs
@@ -128,7 +128,7 @@ mod test {
       "enable_ipv6": false
     }
   },
-  "settings_version": 4
+  "settings_version": 5
 }
 "#;